+++ src/rogallo/commands/ui.py
@@ -41,13 +41,6 @@ class ToggleSidePanel(Command):
 
 
 ##############################################################################
-class StripeLinks(Command):
-    """Toggle the striping of links in the document viewer"""
-
-    BINDING_KEY = "f8"
-
-
-##############################################################################
 class ToggleLinkNumbers(Command):
     """Toggle the display of link numbers in the document viewer"""