+++ src/rogallo/screens/main/screen.py
@@ -71,7 +71,6 @@ from ...commands import (
     SearchHistory,
     SetHome,
     SetHomeToCurrentLocation,
-    StripeLinks,
     ToggleANSIEscapeSequenceHandling,
     ToggleCosyLinkNumbers,
     ToggleLinkNumbers,
@@ -242,7 +241,6 @@ class Main(EnhancedScreen[None]):
         SaveSource,
         SetHome,
         SetHomeToCurrentLocation,
-        StripeLinks,
         ToggleANSIEscapeSequenceHandling,
         ToggleCosyLinkNumbers,
         ToggleLinkNumbers,
@@ -367,7 +365,6 @@ class Main(EnhancedScreen[None]):
                 for host, port in await self._clients.gemini.trust_store.get_hosts()
             ]
             HistorySearchCommands.known_hosts = self._command_line.known_hosts
-        self._viewer.stripe_links = config.stripe_links
         self._viewer.with_link_numbers = config.with_link_jumps
         self._viewer.handle_ansi_escape_sequences = config.handle_ansi_escape_sequences
         self._viewer.cosy_link_numbers = config.cosy_link_jumps
@@ -866,12 +863,6 @@ class Main(EnhancedScreen[None]):
         ):
             self._cache.clear()
             self.notify("All cached content has been cleared.", title="Cache")
-
-    def action_stripe_links_command(self) -> None:
-        """Toggle link striping."""
-        self._viewer.stripe_links = not self._viewer.stripe_links
-        with update_configuration() as config:
-            config.stripe_links = self._viewer.stripe_links
 
     def action_toggle_link_numbers_command(self) -> None:
         """Toggle link numbers."""